requirements.txt
pyusb==1.0.0
requests==2.11.1
websocket-client==0.37.0
